upgrade eslint to version 7.29.0
