+++ package.json
@@ -36,11 +36,7 @@
     "babel-loader": "8.0.2",
     "bluebird": "^3.4.1",
     "copy-webpack-plugin": "4.5.2",
-    "eslint": "^3.2.2",
-    "eslint-config-airbnb": "^10.0.0",
-    "eslint-plugin-import": "^1.12.0",
-    "eslint-plugin-jsx-a11y": "^2.0.1",
-    "eslint-plugin-react": "^6.0.0",
+    "eslint": "^7.29.0",
     "glob": "^7.0.5",
     "html-webpack-plugin": "3.2.0",
     "json-loader": "0.5.7",
@@ -48,8 +44,7 @@
     "shelljs": "^0.7.3",
     "webpack": "4.17.2",
     "webpack-cli": "3.1.0",
-    "webpack-dev-server": "^3.1.14",
-    "webpack-shell-plugin": "0.5.0"
+    "webpack-dev-server": "^3.1.14"
   },
   "author": "WebpackBin",
   "license": "ISC"
