+++ src/popup/Components/MobileIcon.jsx
